+++ include/ftxui/dom/take_any_args.hpp
@@ -5,48 +5,16 @@
 #define FTXUI_DOM_TAKE_ANY_ARGS_HPP
 
 // IWYU pragma: private, include "ftxui/dom/elements.hpp"
-#include <concepts>
+#include <deque>
 #include <ftxui/dom/node.hpp>
 #include <queue>
-#include <ranges>
 #include <stack>
+#include <vector>
 
 namespace ftxui {
 
-template <class T>
-void Merge(Elements& /*container*/, T /*element*/) {}
-
-template <>
 inline void Merge(Elements& container, Element element) {
   container.push_back(std::move(element));
-}
-
-template <class Container>
-concept ElementRange =
-    std::ranges::range<Container> &&
-    std::same_as<std::ranges::range_value_t<Container>, Element>;
-
-template <ElementRange T>
-inline void Merge(Elements& container, T elements) {
-  for (auto& element : elements) {
-    container.push_back(std::move(element));
-  }
-}
-
-template <>
-inline void Merge(Elements& container, std::stack<Element> elements) {
-  while (!elements.empty()) {
-    container.push_back(std::move(elements.top()));
-    elements.pop();
-  }
-}
-
-template <>
-inline void Merge(Elements& container, std::queue<Element> elements) {
-  while (!elements.empty()) {
-    container.push_back(std::move(elements.back()));
-    elements.pop();
-  }
 }
 
 // Turn a set of arguments into a vector.
@@ -57,11 +25,46 @@ Elements unpack(Args... args) {
   return vec;
 }
 
-// Make |container| able to take any number of argments.
+// Make |container| able to take any number of arguments.
 #define TAKE_ANY_ARGS(container)                               \
   template <class... Args>                                     \
-  Element container(Args... children) {                        \
+  inline Element container(Args... children) {                 \
     return container(unpack(std::forward<Args>(children)...)); \
+  }                                                            \
+                                                               \
+  template <class Container>                                   \
+  inline Element container(Container&& children) {             \
+    Elements elements;                                         \
+    for (auto& child : children) {                             \
+      elements.push_back(std::move(child));                    \
+    }                                                          \
+    return container(std::move(elements));                     \
+  }                                                            \
+  template <>                                                  \
+  inline Element container(std::stack<Element>&& children) {   \
+    Elements elements;                                         \
+    while (!children.empty()) {                                \
+      elements.push_back(std::move(children.top()));           \
+      children.pop();                                          \
+    }                                                          \
+    return container(std::move(elements));                     \
+  }                                                            \
+  template <>                                                  \
+  inline Element container(std::queue<Element>&& children) {   \
+    Elements elements;                                         \
+    while (!children.empty()) {                                \
+      elements.push_back(std::move(children.front()));         \
+      children.pop();                                          \
+    }                                                          \
+    return container(std::move(elements));                     \
+  }                                                            \
+  template <>                                                  \
+  inline Element container(std::deque<Element>&& children) {   \
+    Elements elements;                                         \
+    for (auto& child : children) {                             \
+      elements.push_back(std::move(child));                    \
+    }                                                          \
+    return container(std::move(elements));                     \
   }
 
 TAKE_ANY_ARGS(vbox)